// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/models/ApplyDiscountRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ApplyDiscountItemDto } from './ApplyDiscountItemDto';
export type ApplyDiscountRequestDto = {
    jobId: string;
    familyUserId: string;
    code: string;
    items: Array<ApplyDiscountItemDto>;
};