// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/models/ApplyDiscountResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApplyDiscountResponseDto = {
    success?: boolean;
    message?: string | null;
    totalDiscount?: number | string;
    perPlayer?: Record<string, number | string>;
};